requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pytz>=2022.1
python-dotenv>=0.20.0
urllib3>=1.26.0